requirements.txt
dotenv==0.9.9
psycopg2-binary==2.9.10
pydantic>=2.0
pyaml==25.1.0
pymodbus==3.8.6
python-dotenv==1.0.1
PyYAML==6.0.2
